add some refactoring
